Master

See merge request willbes/ProjectWILL!671

+++ application/models/lms/predict/PredictModel.php
@@ -1525,7 +1525,7 @@ class PredictModel extends WB_Model
 
             $addQuery = "";
             if(empty($TakeMockPart) == false){
-                $addQuery = " AND TakeMockPart = ".$TakeMockPart;
+                $addQuery = " AND pg.TakeMockPart = ".$TakeMockPart;
             }
 
             //시험코드
@@ -1589,56 +1589,33 @@ class PredictModel extends WB_Model
                 $result = $query->result_array();
 
                 foreach ($result AS $key => $val) {
-
-                    $column = "
-                        *
-                    ";
-
-                    $from = "
-                        FROM
-                            {$this->_table['predictGradesOrigin']} 
-                    ";
-
-                    $order_by = " ";
-                    $where = " WHERE PrIdx = " . $val['PrIdx'] ." AND ProdCode = ".$val['ProdCode']." AND PpIdx=".$val['PpIdx'];
-                    $query = $this->_conn->query('select ' . $column . $from . $where . $order_by);
-                    $rs = $query->row_array();
-
-                    if(empty($rs)==true){
-                        $orgPoint = $val['OrgPoint'];
-
-                        // 데이터 입력
-                        $data = [
-                            'MemIdx' => $val['MemIdx'],
-                            'PrIdx' => $val['PrIdx'],
-                            'ProdCode' => $val['ProdCode'],
-                            'PpIdx' => $val['PpIdx'],
-                            'OrgPoint' => $orgPoint,
-                            'TakeMockPart' => $val['TakeMockPart'],
-                            'TakeArea' => $val['TakeArea']
-                        ];
-
-                        if ($this->_conn->set($data)->insert($this->_table['predictGradesOrigin']) === false) {
-                            throw new \Exception('시험데이터가 없습니다.');
-                        }
-
+                    if(empty($TakeMockPart) == false) {
+                        $where = ['PrIdx' => $val['PrIdx'], 'PpIdx' => $val['PpIdx'], 'ProdCode' => $val['ProdCode'], 'TakeMockPart' => $TakeMockPart];
                     } else {
-                        $orgPoint = $val['OrgPoint'];
-
-                        // 데이터 입력
-                        $data = [
-                            'MemIdx' => $val['MemIdx'],
-                            'ProdCode' => $val['ProdCode'],
-                            'PpIdx' => $val['PpIdx'],
-                            'OrgPoint' => $orgPoint,
-                            'TakeMockPart' => $val['TakeMockPart'],
-                            'TakeArea' => $val['TakeArea']
-                        ];
-
-                        $this->_conn->set($data)->where('PrIdx', $val['PrIdx']);
-                        if ($this->_conn->update($this->_table['predictGradesOrigin']) === false) {
-                            throw new \Exception('데이터 수정에 실패했습니다.');
+                        $where = ['PrIdx' => $val['PrIdx'], 'PpIdx' => $val['PpIdx'], 'ProdCode' => $val['ProdCode']];
+                    }
+                    try {
+                        if($this->_conn->delete($this->_table['predictGradesOrigin'], $where) === false){
+                            throw new \Exception('삭제에 실패했습니다.');
                         }
+                    } catch (\Exception $e) {
+                        return error_result($e);
+                    }
+
+                    $orgPoint = $val['OrgPoint'];
+                    // 데이터 입력
+                    $data = [
+                        'MemIdx' => $val['MemIdx'],
+                        'PrIdx' => $val['PrIdx'],
+                        'ProdCode' => $val['ProdCode'],
+                        'PpIdx' => $val['PpIdx'],
+                        'OrgPoint' => $orgPoint,
+                        'TakeMockPart' => $val['TakeMockPart'],
+                        'TakeArea' => $val['TakeArea']
+                    ];
+
+                    if ($this->_conn->set($data)->insert($this->_table['predictGradesOrigin']) === false) {
+                        throw new \Exception('시험데이터가 없습니다.');
                     }
 
                 }
@@ -1667,12 +1644,22 @@ class PredictModel extends WB_Model
                 throw new \Exception('합격예측상품 미등록 상태입니다.');
             }
 
-            $this->_conn->where(['ProdCode' => $ProdCode]);
+            $addQuery = "";
+            if(empty($TakeMockPart) == false){
+                $addQuery = " AND pg.TakeMockPart = " . $TakeMockPart;
+                $this->_conn->where(['ProdCode' => $ProdCode, 'TakeMockPart' => $TakeMockPart]);
 
-            if ($this->_conn->delete($this->_table['predictGrades']) === false) {
-                throw new \Exception('성적 삭제에 실패했습니다.');
+                if ($this->_conn->delete($this->_table['predictGrades']) === false) {
+                    throw new \Exception('성적 삭제에 실패했습니다.');
+                }
+            } else {
+                $this->_conn->where(['ProdCode' => $ProdCode]);
+
+                if ($this->_conn->delete($this->_table['predictGrades']) === false) {
+                    throw new \Exception('성적 삭제에 실패했습니다.');
+                }
             }
-
+            
             // 데이터 입력
             if ($mode == 'web') {
                 $data = [
@@ -1693,11 +1680,6 @@ class PredictModel extends WB_Model
                 throw new \Exception('로그생성실패.');
             }
 
-            $addQuery = "";
-            if(empty($TakeMockPart) == false){
-                $addQuery = " AND TakeMockPart = " . $TakeMockPart;
-            }
-
             //시험코드
             $column = "
                 pr.TakeMockPart, pr.TakeArea, pg.PpIdx, pp.Type
@@ -1715,6 +1697,7 @@ class PredictModel extends WB_Model
                           ORDER BY TakeMockPart, TakeArea, pg.PpIdx";
 
             $where = " WHERE pg.ProdCode = " . $ProdCode . $addQuery;
+            //echo "<pre>". 'select' . $column . $from . $where . $order_by . "</pre>";
 
             $query = $this->_conn->query('select ' . $column . $from . $where . $order_by);
